+++ test/Unit/Location/ConstantLocationTest.php
@@ -35,4 +35,32 @@ final class ConstantLocationTest extends Framework\TestCase
 
         self::assertSame($constantName, $location->constantName());
     }
+
+    public function testEqualsReturnsFalseWhenTypesAreDifferent(): void
+    {
+        $one = Location\ConstantLocation::create(Name\ConstantName::fromString('FOO'));
+        $two = $this->createStub(Location\Location::class);
+
+        self::assertFalse($one->equals($two));
+    }
+
+    public function testEqualsReturnsFalseWhenConstantNamesAreDifferent(): void
+    {
+        $faker = self::faker()->unique();
+
+        $one = Location\ConstantLocation::create(Name\ConstantName::fromString($faker->word()));
+        $two = Location\ConstantLocation::create(Name\ConstantName::fromString($faker->word()));
+
+        self::assertFalse($one->equals($two));
+    }
+
+    public function testEqualsReturnsTrueWhenConstantNamesAreEqual(): void
+    {
+        $value = self::faker()->word();
+
+        $one = Location\ConstantLocation::create(Name\ConstantName::fromString($value));
+        $two = Location\ConstantLocation::create(Name\ConstantName::fromString($value));
+
+        self::assertTrue($one->equals($two));
+    }
 }